% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addKey}
\alias{addKey}
\title{addKey}
\usage{
addKey(df, cols, separator, id_col_name)
}
\arguments{
\item{df}{- data frame to add the key to}

\item{cols}{- columns to concatenate}

\item{separator}{- string for separating concatenated columns}

\item{id_col_name}{- name of the newly created key column}
}
\value{
result - the original data frame with an additional key column
}
\description{
add a key to a data frame which is a concatenation of existing columns
}
\section{Requirements}{

  plyr (version >= 1.8.1)
}

\section{Sources}{

  \tabular{ll}{
  \tab Ananda Mahto, February 10, 2014 \cr
  \tab \url{http://stackoverflow.com/questions/21682462/concatenate-columns-and-add-them-to-beginning-of-data-frame} \cr
  }
}

\section{Adapted}{

  \tabular{llllllll}{
  \tab 2015-01-22 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2015-01-22 \tab\tab BLC  \tab\tab Initial version \cr
  }
}
\examples{
# create a concatenated key for comparisons
# cols <- c("START_DATE","Park","StationID","SMPL_FRAC_TYPE_NM","MEDIUM")
# addKey(dfResults, cols, "|", "uniqueID")
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

