% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{app}
\alias{app}
\title{app Class}
\format{\preformatted{List of 4
 $ dbfilepathname    : chr "LOCATION OF DATABASE FILE"
 $ dbusername        : chr ""
 $ dbpwd             : chr ""
 $ sqlQueriesFilepath: chr ""
 - attr(*, "class")= chr "app"
}}
\usage{
app
}
\arguments{
\item{apppathname}{- database name & directory path}

\item{username}{- database username (for authentication)}

\item{pwd}{- database password (for authentication)}
}
\description{
app Class
}
\details{
Create a database object whose connection parameters are contained within the object
}
\section{Requirements}{

  RODBC library
  tools library
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-10-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-10-05  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\keyword{datasets}

