% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.is}
\alias{as.is}
\title{as.is}
\usage{
as.is(x)
}
\arguments{
\item{x}{- any value}
}
\value{
x - Returns the value unchanged.
}
\description{
Leave the value unchanged
}
\note{
This function is used to pass into aggregation functions and similar to allow for
  unchanged values to be passed through.
}
\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-05-30  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab \tab 2014-05-30 \tab \tab BLC  \tab \tab Initial version \cr
  }
}
\seealso{
Other helper functions: \code{\link{cleanUp}};
  \code{\link{convert.magic}}; \code{\link{daysBetween}};
  \code{\link{dfDatesBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

