% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateStdValues}
\alias{calculateStdValues}
\title{calculateStdValues}
\usage{
calculateStdValues(depChar, depCharValue, depCharValue2, coef1, coef2, coef3,
  coef4, coef5, coef6, coef7, coef8)
}
\arguments{
\item{depChar}{- NPSTORET dependent characteristic value (CharDependent)}

\item{depCharValue}{- result value of first dependent characteristic (pH, hardness, Ca+Mg)}

\item{depCharValue2}{- result value of second dependent characteristic (temp, water) if present}

\item{coef1:coef8}{- NPSTORET calculation coefficients 1-8}
}
\value{
resultVal - results value for the matched characteristic
}
\description{
Add calc std values to dfResults
}
\section{Requirements}{

 \tabular{l}{
   RODBC library package loaded \cr
   Working directory set for app \cr
   LoadNPSTORET function \cr
   stringr library package loaded \cr
}
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-07-02 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-07-02    \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}},
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

