% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cleanUp}
\alias{cleanUp}
\title{cleanUp}
\usage{
cleanUp(varList, blnDisconnect = TRUE)
}
\arguments{
\item{varList}{- list of vars to remove   list = c('temp1','temp2')
Make the list a character vector (not a vector of names)
To remove ALL items set   list = ls()}

\item{blnDisconnect}{- disconnect the db connection? (True/False)}
}
\description{
Release resources
}
\note{
This function is specific to the NPSTORET_import functions
  in the future it should be made more generic.

For gc()
 verbose   logical; TRUE = garbage collection prints statistics about cons cells
                           & space allocated for vectors.
 reset	   logical; TRUE = values for maximum space used are reset to the current
                           values.
}
\section{Requirements}{

  N/A
}

\section{Sources}{

  \tabular{ll}{
  \tab M. Nel, July 24, 2012 \cr
  \tab \url{http://stackoverflow.com/questions/11624885/remove-multiple-objects-with-rm} \cr
  \tab David Robinson, Jan 11, 2012 \cr
  \tab \url{http://stackoverflow.com/questions/8813753/what-is-the-difference-between-gc-and-rm} \cr
  \tab Chad Birch, March 16, 2009 \cr
  \tab \url{http://stackoverflow.com/questions/652136/how-can-i-remove-an-element-from-a-list} \cr
  \tab Paul Teetor \cr
  \tab \url{https://www.inkling.com/read/r-cookbook-paul-teetor-1st/chapter-5/recipe-5-12} \cr
  \tab Evans, July 12, 2010 \cr
  \tab \url{http://economics.wordpress.com/2010/07/12/remove-null-elements-from-a-list-in-r/} \cr
  }
}

\section{Adapted}{

  \tabular{llll}{
  \tab 2014-05-08  \tab \tab B. Campbell \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1    \tab\tab 2014-05-08  \tab\tab BLC  \tab \tab Initial version \cr
  \tab 0.2    \tab\tab 2014-06-25  \tab\tab BLC  \tab \tab Extended to include variable input & disconnect option \cr
  \tab 0.3    \tab\tab 2014-07-01  \tab\tab BLC   \tab\tab Added existence checks \cr
  \tab 0.4    \tab\tab 2014-11-10  \tab\tab BLC   \tab\tab Updated documentation & removed plyr references since Filter isn't used \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{convert.magic}}; \code{\link{daysBetween}};
  \code{\link{dfDatesBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

