% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compareToStandards}
\alias{compareToStandards}
\title{compareToStandards}
\usage{
compareToStandards(resultValue, stdValue, loHi)
}
\arguments{
\item{resultValue}{- WQ results}

\item{stdValue}{- WQ standard value (either dependent or independent)}

\item{loHi}{- indicator if WQ standard is a low or high standard
                low = results values should not be below the standard
                high = results values should not be above the standard}
}
\value{
compliance - indicator for whether WQ results were in/out of compliance
                -1 = result below WQ standard
                 0 = result in acceptable range
                +1 = result above WQ standard
}
\description{
Compare results to standard values
}
\section{Requirements}{

 none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-07-03 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-07-03    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-16    \tab\tab BLC   \tab\tab Handle resultValues > stdValue (low threshold std) &
                                                        resultValues < stdValue (high threshold std), & Documentation updates \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

