% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert.magic}
\alias{convert.magic}
\title{convert.magic}
\usage{
convert.magic(obj, type, cols)
}
\arguments{
\item{none}{}
}
\description{
Change data frame column types
}
\note{
Used to change from scientific notation to standard for dataframes
}
\section{Sources}{

  \tabular{ll}{
  \tab \url{http://stackoverflow.com/questions/11261399/function-for-converting-dataframe-column-type} \cr
  }
}

\section{Created}{

  \tabular{ll}{
  \tab Mikko June 29, 2012 \cr
  }
}

\section{Adapted}{

  \tabular{llll}{
  \tab 2014-05-30  \tab\tab B. Campbell \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-30 \tab \tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-06-26  \tab\tab BLC  \tab \tab Added cleanup \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{daysBetween}};
  \code{\link{dfDatesBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

