% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{convertToDataTable}
\alias{convertToDataTable}
\title{convertToDataTable}
\usage{
convertToDataTable(df, key_col)
}
\arguments{
\item{df}{- data frame to add the key to}

\item{key_col}{- key columns}
}
\value{
result - the original data frame as a data.table
}
\description{
converts a dataframe to a data.table
}
\section{Requirements}{

  data.table (version >= 1.9.4)
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2015-01-22 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2015-01-22 \tab\tab BLC  \tab\tab Initial version \cr
  }
}
\examples{
# convert data frame to data table & set key
# cols <- c("START_DATE","Park","StationID","SMPL_FRAC_TYPE_NM","MEDIUM","FIELD_LAB")
# convertToDataTable(dfResults, cols)
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}}; \code{\link{daysBetween}};
  \code{\link{dfDatesBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}};
  \code{\link{trim.leading}}; \code{\link{trim.trailing}};
  \code{\link{trim}}
}

