% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{db-class}
\alias{db}
\alias{db-class}
\title{db Class}
\description{
db Class
}
\details{
Creates a database object whose connection parameters are contained within the object.
}
\section{Fields}{

\describe{
\item{\code{dbfilepath}}{- database name & directory path}

\item{\code{user}}{- database username (for authentication)}

\item{\code{pwd}}{- database password (for authentication)}
}}
\section{Requirements}{

R Libraries:
\itemize{
 \item \link[RODBC]{RODBC}
 \item \link[tools]{tools}
}
}

\section{Sources}{

  \tabular{lll}{
  \tab 2004 \tab Friedrich Leisch \cr
  \tab\tab http://www.r-project.org/conferences/useR-2004/Keynotes/Leisch.pdf \cr
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-10-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-10-05  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-07  \tab\tab BLC   \tab\tab Documentation update & removed sqlQUeriesFilepath since SQL queries are integrated \cr
  \tab 0.3   \tab\tab 2014-11-13  \tab\tab BLC   \tab\tab Documentation update \cr
  \tab 0.4   \tab\tab 2014-11-17  \tab\tab BLC   \tab\tab Converted to S4 class & renamed to db vs. app \cr
  }
}
\examples{
## Create the database object (db) and assign the NPSTORET backend filename.
## Paths should be written UNIX style with / vs. \\ .
## So a windows file path like "C:\\db\\dbname.mdb" should be written "C:/db/dbname.mdb".

\dontrun{
npstoret <- db("C:/my_database/NPSTORET_BE.MDB","","")

Connect to your database via a DSN-less connection.

npstoret <- connect(npstoret, dbfilepath, user, pwd)
}
}

