% Generated by roxygen2 (4.0.2): do not edit by hand
\name{depCharResultsLookup}
\alias{depCharResultsLookup}
\title{depCharResultsLookup}
\usage{
depCharResultsLookup(dfDepCharResults, characteristic, startVisitDate,
  endVisitDate, stationID = 0)
}
\arguments{
\item{dfDepCharResults}{- WQ Results dataframe for pH, temp, hardness (Ca+Mg)
(optional, if already generated)}

\item{startVisitDate}{- target visit date (start of range)}

\item{endVisitDate}{- target visit date (end of range), use startVisitDate if looking
for characteristics for one visit date}

\item{stationID}{- specific stationID, default = 0 for ALL stations}
}
\value{
dataframe - includes dependent characteristic values & info for given
                      characteristic (pH, temp, hardness(Ca+Mg))
                      w/in the visit range (startVisitDate, endVisitDate)
                      for the station identified (or ALL stations if 0)
                      may include:  0 rows (no results found)
                                    1 row (usable for calculations)
                                    many rows (multiple visits or stations)
}
\description{
Fetch specific Dependent Characteristic NPSTORET WQ Result
}
\section{Requirements}{

 \tabular{l}{
   RODBC library package loaded \cr
   Working directory set for app \cr
   LoadNPSTORET function \cr
   sqldf library package loaded \cr
}
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-06-26 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-26    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-18    \tab\tab BLC   \tab\tab Added stringsAsFactors=FALSE to avoid
                            factorization of df character strings &
                            subsequent limitation of df to those values \cr
 }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}},
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

