% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfDatesBetween}
\alias{dfDatesBetween}
\title{dfDatesBetween}
\usage{
dfDatesBetween(df, col = "", firstDate, secondDate)
}
\arguments{
\item{df}{- dataframe object}

\item{col}{- name of the column as a string}

\item{date1}{- first date in YYYY-mm-dd format}

\item{date2}{- second date in YYYY-mm-dd format}
}
\value{
df - dataframe with column dates between the 2 dates or -1 if df isn't a dataframe
}
\description{
Returns subset of a dataframe with values between dates for a specific column
}
\section{Sources}{

\tabular{ll}{
\tab beginneR, March 13, 2014 \cr
\tab \url{http://stackoverflow.com/questions/23622338/subset-a-dataframe-between-2-dates-in-r-better-way} \cr
}
}

\section{Adapted}{

 \tabular{llllllll}{
   \tab 2014-06-26  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
   }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab \tab 2014-06-26 \tab \tab BLC  \tab \tab Initial version \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{convert.magic}};
  \code{\link{daysBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

