% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dfExists}
\alias{dfExists}
\title{dfExists}
\usage{
dfExists(df, dfName = "")
}
\arguments{
\item{df}{- dataframe object}

\item{dfName}{- name of the dataframe as a string}
}
\value{
boolean - TRUE if the name represents a dataframe object, FALSE if not
}
\description{
Object exists and is a data frame object
}
\section{Sources}{

\tabular{ll}{
  \tab Sven Hohenstein, December 13, 2013 \cr
  \tab \url{http://stackoverflow.com/questions/20573119/check-if-data-frame-exists} \cr
  }
}

\section{Adapted}{

\tabular{llllllll}{
  \tab 2014-05-05  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1    \tab\tab 2014-06-25  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{convert.magic}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

