% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExtendedDepCharResults}
\alias{getExtendedDepCharResults}
\title{getExtendedDepCharResults}
\usage{
getExtendedDepCharResults()
}
\value{
df - dfDepCharResults plus pH, Hardness, Temp, Hardness_UOM, Temp_UOM, CalcFormula,
       Values for these variables are in separate columns of the result dataframe.
       CalcStdValue is also provided and is based on CalcFormula and the applicable
       standard characteristics
}
\description{
Fetch dependent char results with lookup values for calculating dependent standards, formulae, & calculated std
}
\section{Assumptions}{

  NPSTORET results are available as dfResults
}

\section{Requirements}{

 none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-09-10 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-10    \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

