% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFormulaUsed}
\alias{getFormulaUsed}
\title{getFormulaUsed}
\usage{
getFormulaUsed(dependency, params)
}
\arguments{
\item{dependency}{- values for NPSTORET characteristics upon which the desired characteristic is based}

\item{params}{- coefficients 1-8 as key value pairs from NPSTORET
               list("coef1" = Coefficient1, "coef2" = Coefficient2, ...)}
}
\value{
string - formula used to calculate the dependent standard value
}
\description{
Fetch the formula used to calculate dependent NPSTORET WQ Standard Values
}
\section{Requirements}{

  RODBC library package loaded
  Working directory set for app
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-06-26 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

 \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-26    \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}},
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

