% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHardnessLnVal}
\alias{getHardnessLnVal}
\title{getHardnessLnVal}
\usage{
getHardnessLnVal(hardness)
}
\arguments{
\item{hardness}{- hardness (Ca+Mg) result value for a given stationID & visit (dfResults)}
}
\value{
lnVal - numeric value for the log 10 of hardness OR -99999999 if hardness = 0
}
\description{
Fetch the hardness (Ca+Mg) lnVal for calculating hardness dependent characteristic standards
}
\section{Requirements}{

 R libraries:
  \itemize{
   \item RODBC
}
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-06-30 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-30    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-08    \tab\tab BLC   \tab\tab fixed to be natural log (ln) in R vs log10
                            log(hardness) vs. log(10, hardness) where 10 identifies base 10
                            & added check for hardness values = "*Non-detect"
                            added NULL check & as.numeric for handling hardness log evaluations \cr
  \tab 0.3   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update \cr
 }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

