% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getParkData}
\alias{getParkData}
\title{getParkData}
\usage{
getParkData(df, parkAbbrev)
}
\arguments{
\item{df}{- data frame holding park data}

\item{parkAbbrev}{- park name abrev (ARCH, CANY, etc.)}

\item{col}{- column to compare from dataframe}
}
\description{
Pull park data from dataframe
}
\section{Sources}{

  2014-05-07 B. Campbell 0.1 Initial version
}

\section{Revisions}{

  0.1   2014-05-07  BLC   Initial version
}

