% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_NPSTORET_functions.R
\name{getParksList}
\alias{getParksList}
\title{getParksList}
\usage{
getParksList(df)
}
\arguments{
\item{df}{- data frame holding park data}
}
\description{
Return a list of Park abbreviations from dataframe
}
\section{Requirements}{

  assumes df column name for Park abbreviation is "Park"
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-05-08 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-08  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other NPSTORET functions: \code{\link{getFilteredNPSTORETData}};
  \code{\link{getParkData}};
  \code{\link{loadNPSTORETData}};
  \code{\link{loadNPSTORETWQData}}
}

