% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_WQstds_functions.R
\name{getRAWDepWQStdChars}
\alias{getRAWDepWQStdChars}
\title{getRAWDepWQStdChars}
\usage{
getRAWDepWQStdChars()
}
\description{
fetch ALL NPSTORET WQ Standard Characteristics for Dependent (Uncalculated) Standards
}
\section{Requirements}{

 R libraries:
  \itemize{
   \item RODBC
  }
}

\section{Sources}{

\tabular{llllllll}{
   \tab 2014-06-25 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
   }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-25    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-08    \tab\tab BLC   \tab\tab Drop unused levels \cr
  \tab 0.3   \tab\tab 2014-09-10    \tab\tab BLC   \tab\tab Remove all extra columns except CalcFormula
                            since these are time/date dependent & require lookups
                            Hardness, pH, & Temp will be added to RESULTS vs chars dataframes \cr
  \tab 0.5   \tab\tab 2014-09-18    \tab\tab BLC   \tab\tab Added stringsAsFactors = FALSE to prevent
                            factorization of df strings & subsequent
                            limitation of df values, fix spaces & reordering df \cr
  \tab 0.6   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update, add check for dfRAWStds & remove sqldf depedency \cr
  \tab 0.7   \tab\tab 2015-01-06    \tab\tab BLC   \tab\tab Fix unused argument (IsDependent=0) error \cr
 }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

