% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getStdDepCharDisplayName}
\alias{getStdDepCharDisplayName}
\title{getStdDepCharDisplayName}
\usage{
getStdDepCharDisplayName(depCharacteristic)
}
\arguments{
\item{depCharacteristic}{- CharDependent value from NPSTORET (dfRAWDepChars)}
}
\value{
display_name - string value for the display name of the characteristic used to
         determine the WQ standard
         This includes variables like pH, lnVal (hardness, Ca+Mg), temp
         The values for these variables are in separate columns of the result dataframe.
}
\description{
Fetch the display name for a characteristic's dependent characteristic
}
\section{Requirements}{

 none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-09-17 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-17    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

