% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_WQstds_functions.R
\name{getStdFormula}
\alias{getStdFormula}
\title{getStdFormula}
\usage{
getStdFormula(depCharacteristic)
}
\arguments{
\item{depCharacteristic}{- CharDependent value from NPSTORET (dfRAWDepChars)}
}
\value{
calc - string value for the calculation used to determine the WQ standard
         This includes variables like pH, lnVal (hardness, Ca+Mg), temp, & coefficients 1-8.
         The values for these variables are in separate columns of the result dataframe.
}
\description{
Fetch the formula for calculating dependent characteristic standards
}
\section{Requirements}{

 none
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-06-30 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-30    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

