% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWQRawStdCharacteristics}
\alias{getWQRawStdCharacteristics}
\title{getWQRawStdCharacteristics:}
\usage{
getWQRawStdCharacteristics()
}
\description{
Fetch ALL NPSTORET WQ Standard Characteristics for Dependent (Calculated) &
             Independent (Uncalculated) Standards
}
\note{
Standards are 'raw' pulled directly from NPSTORET - NO CALCULATED VALUES!
      DO NOT USE for dependent characteristic standard values
}
\section{Requirements}{

R libraries:
\itemize{
   \item RODBC
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-06-13 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-13 \tab\tab   BLC \tab\tab  Initial version \cr
  \tab 0.2   \tab\tab 2014-06-24 \tab\tab   BLC  \tab\tab Added StnID column & converted to non-scientific notation \cr
  \tab 0.3   \tab\tab 2014-09-08 \tab\tab   BLC  \tab\tab Drop unused levels \cr
  \tab 0.4   \tab\tab 2014-10-05 \tab\tab   BLC  \tab\tab Add check to ensure dfRAWStds is a data.frame \cr
  \tab 0.5   \tab\tab 2014-11-11 \tab\tab   BLC  \tab\tab Documentation update \cr
  \tab 0.6   \tab\tab 2014-12-31 \tab\tab   BLC  \tab\tab Fix dfRAWstds typo \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

