% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWQStdValues}
\alias{getWQStdValues}
\title{getWQStdValues}
\usage{
getWQStdValues(dfWQstds)
}
\arguments{
\item{dependency}{- value for NPSTORET characteristic upon which the desired characteristic is based}
}
\description{
Fetch NPSTORET WQ Standard Values for Calculated & Uncalculated Characteristics
}
\section{Requirements}{

 \tabular{l}{
   RODBC library package loaded \cr
   Working directory set for app \cr
}
}

\section{Sources}{

  Reference NPSTORET reportUtilities module, CalcDepStandard function
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 2014-05-21 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  \tab 2014-06-05 \tab\tab B. Campbell \tab\tab 0.2 \tab\tab Adjusted to mirror NPSTORET calculations \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{lookupDependentCharResult}},
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

