% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeData}
\alias{initializeData}
\title{initializeData}
\usage{
initializeData()
}
\arguments{
\item{app}{- The application variable in the Global Environment which contains the
             database filename and path (app[['dbfilepathname']]).
             Without this value the function will fail to find the defined file name and path and
             the database connections for loading from NPSTORET will fail
             with a somewhat cryptic "first argument is not an open RODBC channel" error.}
}
\description{
Load NPSTORET data into global (working) environment
             The following data frames are generated:
             \itemize{
             \item dfResults - contains ALL NPSTORET results
             \item dfRAWStds - unmodified NPSTORET standard characteristics
             \item dfIndepChars - unmodified NPSTORET independent characteristics
             \item dfRAWDepChars - unmodified NPSTORET dependent characteristics
             \item dfDepCharResults - sample results for pH, hardness (Ca+Mg), temperature (dependencies)
             \item dfRAWChars - all unmodified NPSTORET characteristics
             \item depStds - dependent standards
             \item indepStds - independent standards
             \item dStds - dependent standards
             \item iStds -  independent standards
             \item pH - all pH sample results
             \item Hardness - all Hardness (Ca + Mg) sample results
             \item Temp - all water Temperature sample results
             }
             * Dependent characteristics include standard values which depend on other sample results (like pH, temp, hardness (Ca + Mg), time).
               Independent characteristics include standard values that do not rely on other sample results.
}
\section{Requirements}{

  \tabular{l}{
  RODBC library \cr
  sqldf library \cr
  app dbfilepathname for valid NPSTORET database with required tables \cr
  }
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-10-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-10-05  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other app initialization functions: \code{\link{initializeEnvironment}}
}

