% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeEnvironment}
\alias{initializeEnvironment}
\title{initializeEnvironment}
\usage{
initializeEnvironment()
}
\arguments{
\item{none}{}
}
\description{
Load functions and information into global (working) environment
}
\section{Requirements}{

  none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-09-30 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-30  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other app initialization functions: \code{\link{initializeData}}
}

