% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert.column}
\alias{insert.column}
\title{insert.column}
\usage{
insert.column(df, col, afterCol, copyCol, blnSciNotation = FALSE)
}
\arguments{
\item{df}{- dataframe to insert column into}

\item{col}{- name of column to insert}

\item{afterCol}{- column name before where the inserted column will be placed}

\item{copyCol}{- copied column name (default = '' meaning no column is copied)}

\item{blnSciNotation}{- whether column should be in scientific notiation or not (default = FALSE)}
}
\value{
dataframe w/ new column
}
\description{
Insert column into dataframe
}
\note{
Used to insert a column into the middle of a dataframe
}
\section{Sources}{

  \tabular{ll}{
  \tab Peter McMahan, August 2, 2009 \cr
  \tab \url{http://stackoverflow.com/questions/1177919/does-column-exist-and-how-to-rearrange-columns-in-r-data-frame} \cr
  }
}

\section{Adapted}{

  \tabular{llllllll}{
  \tab 2014-06-25  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab \tab 2014-06-25 \tab \tab BLC  \tab \tab Initial version \cr
  \tab 0.2    \tab\tab 2014-06-26  \tab\tab BLC   \tab\tab Added cleanup \cr
  \tab 0.3    \tab\tab 2014-11-11  \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{convert.magic}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

