% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{is.not.null}
\alias{is.not.null}
\title{is.not.null}
\usage{
is.not.null(x)
}
\arguments{
\item{x}{- Value to check}
}
\value{
TRUE - if value is not NULL

FALSE - if value is NULL
}
\description{
Determine if a value is or is not null
}
\section{Sources}{

\tabular{ll}{
  \tab \url{http://stackoverflow.com/questions/2175809/alternative-to-is-null-in-r Etiennebr, 2/2/2010} \cr
  }
}

\section{Adapted}{

 \tabular{llllllll}{
  \tab 2014-05-08  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1    \tab\tab 2014-05-08  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2    \tab\tab 2014-11-10  \tab\tab BLC   \tab\tab Updated documentation \cr
  }
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}};
  \code{\link{trim.leading}}; \code{\link{trim.trailing}};
  \code{\link{trim}}
}

