% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{isDateBetween}
\alias{isDateBetween}
\title{isDateBetween}
\usage{
isDateBetween(firstDate, secondDate, evalDate, blnInclusive = TRUE)
}
\arguments{
\item{firstDate}{- first date in YYYY-mm-dd format}

\item{secondDate}{- second date in YYYY-mm-dd format}

\item{evalDate}{- date to evaluate in YYYY-mm-dd format}

\item{blnInclusive}{- boolean, include first & second dates in the span
                 (i.e. evalDate is OK if it equals either date)}
}
\value{
boolean - TRUE if evalDate is 'between' (or equal, if inclusive=TRUE)
            date1 and date2, FALSE if evalDate outside the date range
}
\description{
Returns TRUE or FALSE depending upon if the Date lies between start & end dates
}
\section{Sources}{

N/A
}

\section{Adapted}{

\tabular{llllllll}{
  \tab 2014-06-26  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1    \tab\tab 2014-06-26  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2    \tab\tab 2014-11-10  \tab\tab BLC   \tab\tab Update documentation \cr
  }
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{iterateDataLoad}};
  \code{\link{mgsub}}; \code{\link{pkgTest}};
  \code{\link{readSQL}}; \code{\link{trim.leading}};
  \code{\link{trim.trailing}}; \code{\link{trim}}
}

