% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{iterateDataLoad}
\alias{iterateDataLoad}
\title{iterateDataLoad}
\usage{
iterateDataLoad(dfName, tableName, maxResults, numRowsAtTime)
}
\arguments{
\item{dfName}{= new data frame name}

\item{tableName}{= database name of table or query (no function columns in data unless R code re-creates that function)}

\item{maxResults}{= maximum number of results to return, 0 = ALL results}

\item{numRowsAtTime}{= number of results records to return at a time}
}
\value{
progress bar until the data is loaded
}
\description{
Iterate through an ODBC load from tables or query and populates a data frame (dfName)
                 displays a progress bar while data frame is loaded
}
\section{Sources}{

\tabular{ll}{
 \tab \url{http://www.inside-r.org/r-doc/utils/winProgressBar} \cr
 \tab \url{http://rgm3.lab.nig.ac.jp/RGM/R_rdfile?f=utils/man/windows/winProgressBar.Rd&d=R_Rel} \cr
 }
}

\section{Adapted}{

\tabular{llllllll}{
  \tab 2014-05-05 \tab \tab B. Campbell \tab \tab 0.1 \tab \tab Initial version \cr
  }
}

\section{Revisions}{

 \tabular{llllllll}{
  \tab 0.1   \tab \tab 2014-05-05 \tab \tab BLC  \tab \tab Initial version \cr
  \tab 0.2   \tab \tab 2014-06-26 \tab \tab BLC  \tab \tab Added cleanup \cr
  }
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{mgsub}}; \code{\link{pkgTest}};
  \code{\link{readSQL}}; \code{\link{trim.leading}};
  \code{\link{trim.trailing}}; \code{\link{trim}}
}

