% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadNPSTORETData}
\alias{loadNPSTORETData}
\title{loadNPSTORETData
Load NPSTORET data & populate data frame (dfName)}
\usage{
loadNPSTORETData()
}
\description{
loadNPSTORETData
Load NPSTORET data & populate data frame (dfName)
}
\section{Requirements}{

R libraries:
\itemize{
  \item RODBC
}
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-05-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-05  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-06-24  \tab\tab BLC   \tab\tab Updated from SQLquery_NPSTORET_qryResultsAllOrganizations_altered.txt to SQLquery_NPSTORET.txt
                             (changes filename only, not SQL) \cr
  \tab 0.3   \tab\tab 2014-06-26  \tab\tab BLC   \tab\tab Added cleanup \cr
  \tab 0.4   \tab\tab 2014-09-18  \tab\tab BLC   \tab\tab Added stringsAsFactors = FALSE to prevent factorization
                            of character strings & subsequent limitation of resulting dfs
                            to values in original column's factor list
                            Drop levels to avoid retaining unused levels in df \cr
  \tab 0.5   \tab\tab 2014-09-19  \tab\tab BLC  \tab\tab Added odbcClose for connection cleanup \cr
  \tab 0.6   \tab\tab 2014-10-09  \tab\tab BLC  \tab\tab Documentation update \cr
  \tab 0.7   \tab\tab 2014-10-13  \tab\tab BLC  \tab\tab Revised connection to app[["connection"]] vs internal connection \cr
  \tab 0.8   \tab\tab 2014-11-11  \tab\tab BLC  \tab\tab Updated documentation & removed odbcClose \cr
  \tab 0.9   \tab\tab 2014-11-12  \tab\tab BLC   \tab\tab Revise SQL paths to sqlPath \cr
  \tab 0.10   \tab\tab 2014-12-30  \tab\tab BLC   \tab\tab Update dbConn to use npstoret@connect object connection \cr
  }
}
\seealso{
Other NPSTORET functions: \code{\link{getFilteredNPSTORETData}};
  \code{\link{getParkData}}; \code{\link{getParksList}};
  \code{\link{loadNPSTORETWQData}}
}

