% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_NPSTORET_functions.R
\name{loadNPSTORETWQData}
\alias{loadNPSTORETWQData}
\title{loadNPSTORETWQData}
\usage{
loadNPSTORETWQData(datatype = "Results", filter = "")
}
\arguments{
\item{datatype}{- String (enclosed in quotes) that identifies what data should be retrieved.
                This identifies which SQL query should be run to retrieve data from NPSTORET.
                Possible values, what they return, & the related SQL query include:
                \itemize{
                 \item \code{Result} - all NPSTORET results (\href{/doc/SQLquery_NPSTORET_WQResults.txt}{SQLquery_NPSTORET_WQResults})
                 \item \code{Characteristics} - all WQ characteristics (SQLquery_NPSTORET_Characteristics.txt)
                 \item \code{WQProjStnStds} - all WQ locations tied to assigned projects & standards (SQLquery_NPSTORET_WQSP.txt)
                 \item \code{WQProjStnStdCrit} - all WQ locations tied to assigned projects, standards, & criteria (SQLquery_NPSTORET_WQSPC.txt)
                 \item \code{WQProjStnStdCritAlt} - all WQ locations tied to assigned projects, standards, & criteria (SQLquery_NPSTORET_WQProjStnCritStds.txt)
                 \item \code{WQLocales} - all WQ locations tied to assigned projects & standards (SQLquery_NPSTORET_WQLocales.txt)
                 \item \code{WQDepChars} - all WQ dependent characteristics (SQLquery_NPSTORET_WQResults_DepChars.txt)
                 \item \code{WQDepCharsPivot} - all WQ dependent characteristics pivoted to wide format (vs. long) (SQLquery_NPSTORET_WQResults_DepChars_Pivot.txt)
                 \item \code{WQIndepChars} - all WQ standard criteria tied to characteristics (SQLquery_NPSTORET_WQCriteria.txt)
                 \item \code{WQstds} - all WQ standards (SQLquery_NPSTORET_WQStandards.txt)
                 \item \code{WQstdsFiltered} - all WQ standard criteria (filtered by StandardID) (SQLquery_NPSTORET_WQStandardsFiltered.txt)
                 \item \code{WQCriteria} - all WQ standard criteria (SQLquery_NPSTORET_WQCriteria.txt)
                }

                \emph{Time required to retrieve a data set depends on the size of the data sets being retrieved
                and the computer's CPU.

                A result set of ~455,000 with 66 fields takes approximately 1 minute to load the data frame.

                Good time to take a break from the computer screen if you have a very large data set!
                }}

\item{filter}{- the WQ standard to filter by (see examples below)}
}
\value{
df - dataframe with the desired NPSTORET results, characteristics, criteria (as selected by "dataType")
}
\description{
Load NPSTORET data & populate data frame
}
\section{Requirements}{

r Libraries:
\itemize{
 \item \code{\link[RODBC]{RODBC}}
}
}

\section{Notes}{

  SQL queries for NPSTORET data are contained within the npstoret2R package

 ------------------------------------------------------------------------------------------------------
      \code{default.stringsAsFactors()} is typically TRUE and limits any character strings to the factors in the
      original df, so different values not in the factor cannot be entered.
      \code{stringsAsFactors} can ONLY be set at the BEGINNING of the df when it is created
      if a df is created from another df, that df must have been set with \code{stringsAsFactors = FALSE}
      to avoid the issue.
      So, the first df must set \code{stringsAsFactors = FALSE} when its created
      otherwise character strings for a data frame field cannot be changed to values other than
      those that initially exist within the data frame.
 ------------------------------------------------------------------------------------------------------
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-05-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-05  \tab\tab BLC  \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-05-20  \tab\tab BLC  \tab\tab Modified loadNPSTORETData() to include
                         datatype to ID SQL query to load &
                         filter to identify the subset of records to load \cr
  \tab 0.3   \tab\tab 2014-06-02  \tab\tab BLC  \tab\tab Revised Results to limit results fields to speed up query
                         replaced "SQLquery_NPSTORET.txt" with "SQLquery_NPSTORET_WQResults.txt"
                         Added Characteristics to pull the full tblCharacteristics from NPSTORET
                          for tying tblResults to tblWQStandardCriteria \cr
  \tab 0.4   \tab\tab 2014-06-04  \tab\tab BLC  \tab\tab Added locales to bring in WQ sampling stations & their related
                         projects, parks, and WQ assignments \cr
  \tab 0.5   \tab\tab 2014-06-26  \tab\tab BLC  \tab\tab Added cleanup \cr
  \tab 0.6   \tab\tab 2014-09-10  \tab\tab BLC  \tab\tab Added stringsAsFactors = FALSE to enable dfs to *not* automatically
                         convert strings -> factors (default.stringsAsFactors() is TRUE)
                         which can cause later manipulation problems \cr
  \tab 0.7   \tab\tab 2014-09-19  \tab\tab BLC  \tab\tab Added odbcClose for connection cleanup \cr
  \tab 0.8   \tab\tab 2014-10-05  \tab\tab BLC  \tab\tab Updated directory location for package \cr
  \tab 0.9   \tab\tab 2014-10-13  \tab\tab BLC  \tab\tab Revised connection to app[["connection"]] vs internal connection \cr
  \tab 0.10   \tab\tab 2014-11-11  \tab\tab BLC  \tab\tab Documentation updated & removed odbcClose to leave connection available \cr
  \tab 0.11   \tab\tab 2014-11-12  \tab\tab BLC   \tab\tab Revise SQL paths to sqlPath \cr
  \tab 0.12   \tab\tab 2014-12-30  \tab\tab BLC   \tab\tab Update dbConn to use npstoret@connect object connection \cr
  }
}
\examples{
## Typically calls are made to retrieve all results, then WQ characteristics.

    \dontrun{
       # get results
       dfResults <- loadNPSTORETData("Results")

       # get characteristics related to locations, projects, & criteria
       dfChars <- loadNPSTORETData("WQProjStnStdCritAlt")
    }

## Limited filtering is available for water quality standards.
  \dontrun{
       # get a filtered list of standards for Colorado Cold Water
       dfCOCWStds <- loadNPSTORETData("WQstdsFiltered","CO_ALCW1")
    }
}
\seealso{
Other NPSTORET functions: \code{\link{getFilteredNPSTORETData}};
  \code{\link{getParkData}}; \code{\link{getParksList}};
  \code{\link{loadNPSTORETData}}
}

