% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_WQ_results_functions.R
\name{lookupDependentCharResult2}
\alias{lookupDependentCharResult2}
\title{lookupDependentCharResult2}
\usage{
lookupDependentCharResult2(depChar, park, stationID, startDate, startTimeZone,
  smpl_frac, medium, field_lab, uom)
}
\arguments{
\item{depChar}{- NPSTORET dependent characteristic value (pH, Temp, Hardness)}

\item{park}{- NPSTORET Park}

\item{stationID}{- NPSTORET StationID}

\item{startDate}{- NPSTORET visit StartDate}

\item{startTimeZone}{- NPSTORET visit StartTimeZone}

\item{smpl_frac}{- NPSTORET sample fraction type name}

\item{medium}{- NPSTORET measurement medium (MEDIUM)}

\item{field_lab}{- NPSTORET field or lab mesurement (FIELD_LAB)}

\item{uom}{- NPSTORET unit of measure (UOM)#'}
}
\value{
resultVal - results value for the matched characteristic
}
\description{
Lookup dependent characteristic result value (pH, H20 temp, hardness)
}
\section{Requirements}{

  \itemize{
  \item \link[RODBC]{RODBC} library package loaded
  \item Working directory set for app
  \item LoadNPSTORET function
  \item \[stringr]{stringr} library package loaded
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-07-02 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
  \tabular{ll}{
  \tab Richie Cotton, April 24, 2012 \cr
  \tab \url{http://stackoverflow.com/questions/10294284/remove-all-special-characters-from-a-string-in-r} \cr
  \tab Dason, Jun 2, 2012 \cr
  \tab \url{http://stackoverflow.com/questions/10865095/why-do-i-get-warning-longer-object-length-is-not-a-multiple-of-shorter-object-l} \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-07-02    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-11    \tab\tab BLC   \tab\tab Fixed match to return proper result value
                            Replaced length with nrow to get proper # of rows (vs. df width/cols)
                            Replaced == with \%in\% for comparisons \cr
  \tab 0.3   \tab\tab 2014-09-25    \tab\tab BLC   \tab\tab Park, sample fraction type name added to parameters \cr
  \tab 0.4   \tab\tab 2014-09-25    \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other WQ Results functions: \code{\link{AddResultColumns}}
}

