% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookupIndependentCharStd}
\alias{lookupIndependentCharStd}
\title{lookupIndependentCharStd}
\usage{
lookupIndependentCharStd(LocChDefNum, projectID, stationID, uom, medium,
  field_lab)
}
\arguments{
\item{LocChDefNum}{- WQ result local independent characteristic ID number (LocChDef_IS_NUMBER)}

\item{projectID}{- WQ result ProjectID}

\item{stationID}{- WQ result StationID}

\item{uom}{- WQ result unit of measure (UOM)}

\item{medium}{- WQ result measurement medium (MEDIUM)}

\item{field_lab}{- WQ result field or lab mesurement (FIELD_LAB)}
}
\value{
indepCharStdValue - standard value for the matched independent characteristic
                      at the station/project identified
}
\description{
Lookup independent characteristic std value
}
\section{Requirements}{

none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-07-02 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
  \tabular{l}{
  Richie Cotton, April 24, 2012 \cr
  \url{http://stackoverflow.com/questions/10294284/remove-all-special-characters-from-a-string-in-r} \cr
  }
}

\section{Adapted}{

  -
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-07-02    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-07-08    \tab\tab BLC   \tab\tab Removed with & identified dfIndepChars for df variables \cr
  \tab 0.3   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}}
}

