% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{mgsub}
\alias{mgsub}
\title{mgsub}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{- string to search for}

\item{replacement}{- string to replace search pattern with}

\item{x}{- string to inspect}

\item{...}{- additional gsub parameters}
}
\value{
result - the original x string with pattern replaced by replacement
}
\description{
gsub function for multiple replacements
}
\section{Sources}{

  \tabular{ll}{
  \tab Theodore Lytras, March 6, 2013 \cr
  \tab \url{http://stackoverflow.com/questions/15253954/replace-multiple-arguments-with-gsub} \cr
  }
}

\section{Adapted}{

  \tabular{llllllll}{
  \tab 2014-09-30 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-30 \tab\tab BLC  \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-11 \tab\tab BLC  \tab\tab Updated documentation \cr
  }
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{pkgTest}};
  \code{\link{readSQL}}; \code{\link{trim.leading}};
  \code{\link{trim.trailing}}; \code{\link{trim}}
}

