% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgTest}
\alias{pkgTest}
\title{pkgTest}
\usage{
pkgTest(x)
}
\arguments{
\item{x}{- String name of the desired package.}
}
\description{
Ensure required packages are installed, if not install them
}
\section{Sources}{

 \tabular{ll}{
 \tab \url{http://stackoverflow.com/questions/9341635/how-can-i-check-for-installed-r-packages-before-running-install-packages} \cr
 }
}

\section{Adapted}{

  \tabular{llllllll}{
  2014-05-05 \tab \tab B. Campbell \tab \tab 0.1 \tab \tab Initial version \cr
  }
}

\section{Revisions}{

 \tabular{llllllll}{
   0.1   \tab \tab 2014-05-05 \tab \tab BLC  \tab \tab Initial version \cr
   0.2   \tab \tab 2014-06-26 \tab \tab BLC  \tab \tab Added cleanup \cr
   0.3   \tab \tab 2015-01-15 \tab \tab BLC  \tab \tab Documentation update \cr
 }
}
\examples{
# Test for the dplyr package and install it if necessary.
# pkgTest("dplyr")
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{readSQL}}
}

