% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readSQL}
\alias{readSQL}
\title{readSQL}
\usage{
readSQL(path)
}
\arguments{
\item{path}{= SQL file location}
}
\value{
sql - Returns the SQL string from the SQL file.
}
\description{
Read SQL queries from separate SQL file
}
\note{
Long SQL queries become unreadable from R so it's best to keep
       them separate, this function reads & returns the SQL from the query file.
       To bind parameters into the queries, use /%s where the parameter will go
       in the .sql file, then add in the parameters in R using sprintf.
       Include the .sql file by using source(result of readSQL)
}
\section{Sources}{

\tabular{ll}{
 \tab \url{http://stackoverflow.com/questions/13384555/string-continuation-across-multiple-lines-no-newline-characters/13384777#'13384777} \cr
 \tab \url{http://stackoverflow.com/questions/8726882/how-is-it-possible-in-r-to-include-external-files-with-source-code} \cr
 }
}

\section{Adapted}{

  \tabular{llllllll}{
  \tab 2014-05-05 \tab \tab B. Campbell \tab \tab 0.1 \tab \tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1 \tab \tab 2014-05-05 \tab \tab BLC  \tab \tab Initial version \cr
  \tab 0.2 \tab \tab 2014-11-10 \tab \tab BLC  \tab \tab Updated documentation \cr
  }
}
\seealso{
Other helper functions: \code{\link{as.is}};
  \code{\link{cleanUp}}; \code{\link{convert.magic}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}
}

