% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_db_class.R
\name{RODBC-class}
\alias{RODBC-class}
\title{RODBC}
\description{
RODBC
}
\details{
Handles ODBC database connections to R.
}
\section{Requirements}{

R Libraries:
\itemize{
 \item \link[RODBC]{RODBC}
}
}

\section{References}{

  \tabular{lll}{
  \tab Jan. 30, 2012 \tab Kyle Brandt \cr
  \tab\tab http://stackoverflow.com/questions/9067492/example-of-using-an-s3-class-in-a-s4-object \cr
  \tab Oct. 14, 2011 \tab Triad Sou. \cr
  \tab\tab http://stackoverflow.com/questions/7758748/documenting-setas-and-setoldclass-with-roxygen \cr
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-11-24 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-11-24  \tab\tab BLC   \tab\tab Added to resolve S4 connect slot error \cr
  }
}

