% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setDbFile}
\alias{setDbFile}
\title{setDbFile}
\usage{
setDbFile(dbObject, dbfile)
}
\arguments{
\item{dbObject}{- database object}

\item{dbfile}{- database name & directory path}
}
\description{
setDbFile
}
\details{
Set the database file name and path.
}
\section{Requirements}{

R Libraries:
\itemize{
 \item \link[RODBC]{RODBC}
 \item \link[tools]{tools}
}
}

\section{References}{

  \tabular{lll}{
  \tab 2004 \tab Friedrich Leisch \cr
  \tab\tab http://www.r-project.org/conferences/useR-2004/Keynotes/Leisch.pdf \cr
  \tab May 6, 2014 \tab Rappster \cr
  \tab\tab http://stackoverflow.com/questions/23495627/roxygen2-s4-generic-functions-are-not-exported-unless-a-method-is-also-defined \cr
  \tab March 11, 2014 \tab Hadley Wickham \cr
  \tab\tab https://groups.google.com/forum/#!topic/rdevtools/sq3DG0oj058 \cr
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-11-18 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-11-18  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\examples{
\dontrun{
  #set database file & path
  setDbFile(db,"C:/mydatabase/dbfile.accb")
}
}
\seealso{
Other Application settings: \code{\link{connect}};
  \code{\link{db-class}}; \code{\link{setDbPwd}};
  \code{\link{setDbUser}}
}

