% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_db_class.R
\name{setDbUser}
\alias{setDbUser}
\title{setDbUser}
\usage{
setDbUser(dbObject, user)
}
\arguments{
\item{dbObject}{- database object}

\item{user}{- database username (for authentication)}
}
\description{
setDbUser
}
\details{
Sets database username
}
\section{Requirements}{

R Libraries:
\itemize{
 \item \link[RODBC]{RODBC}
 \item \link[tools]{tools}
}
}

\section{References}{

  \tabular{lll}{
  \tab 2004 \tab Friedrich Leisch \cr
  \tab\tab http://www.r-project.org/conferences/useR-2004/Keynotes/Leisch.pdf \cr
  \tab May 6, 2014 \tab Rappster \cr
  \tab\tab http://stackoverflow.com/questions/23495627/roxygen2-s4-generic-functions-are-not-exported-unless-a-method-is-also-defined \cr
  \tab March 11, 2014 \tab Hadley Wickham \cr
  \tab\tab https://groups.google.com/forum/#!topic/rdevtools/sq3DG0oj058 \cr
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-11-18 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-11-18  \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\examples{
\dontrun{
 setDbUser(db,"myusername")
}
}
\seealso{
Other Application settings: \code{\link{connect}};
  \code{\link{db-class}}; \code{\link{setDbFile}};
  \code{\link{setDbPwd}}
}

