% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_utility_functions.R
\name{trim.leading}
\alias{trim.leading}
\title{trim.leading}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{- string to remove whitespace from}
}
\value{
result - x w/o leading whitespace (string)
}
\description{
remove leading whitespace from string
}
\section{Requirements}{

  N/A
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2010-02-14 \tab\tab f3lix \tab\tab 0.1 \tab\tab Initial version \cr
  \tab \url{http://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace-in-r} \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2015-03-24 \tab\tab BLC  \tab\tab Initial version \cr
  }
}
\examples{
#
# trim.leading(tResults$RESULT_DATA)
#
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{daysBetween}}; \code{\link{dfDatesBetween}};
  \code{\link{dfExists}}; \code{\link{insert.column}};
  \code{\link{is.not.null}}; \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}};
  \code{\link{trim.trailing}}; \code{\link{trim}}
}

