% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AddResultColumns}
\alias{AddResultColumns}
\title{AddResultColumns}
\usage{
AddResultColumns(results)
}
\arguments{
\item{results}{- result dataframe}
}
\value{
df - dataframe with columns added
}
\description{
Adds standards column to result dataframe
              including dependent characteristics result values
              (pH, Hardness, Temp), their units of measure (UOM),
              calculation coefficients, formulae as well
              as standards values and compliance information
              Values are initially set as NA until populated
              by sample (pH, Hardness, Temp) & standard values
}
\section{Requirements}{

  \tabular{l}{
  RODBC library package loaded \cr
  Working directory set for app \cr
  LoadNPSTORET function \cr
  }
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-09-25 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-25    \tab\tab BLC   \tab\tab Initial version \cr
  }
}
\seealso{
Other WQ Results functions: \code{\link{lookupDependentCharResult2}}
}

