% Generated by roxygen2 (4.0.2): do not edit by hand
\name{daysBetween}
\alias{daysBetween}
\title{daysBetween}
\usage{
daysBetween(firstDate, secondDate)
}
\arguments{
\item{firstDate}{- first date to compare}

\item{secondDate}{- second date to compare}
}
\value{
integer - # of days between the 2 dates, 0 if first & second dates are the same
                  or -1 if values aren't dates
}
\description{
Returns number of days between two dates
}
\section{Sources}{

  N/A
}

\section{Adapted}{

  \tabular{llllllll}{
  \tab 2014-06-26  \tab\tab B. Campbell  \tab\tab 0.1  \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
   \tab 0.1    \tab\tab 2014-06-26  \tab\tab BLC   \tab\tab Initial version \cr
   \tab 0.2    \tab\tab 2014-11-10  \tab\tab BLC   \tab\tab Updated documentation  \cr
   }
}
\seealso{
Other helper functions: \code{\link{addKey}};
  \code{\link{as.is}}; \code{\link{cleanUp}};
  \code{\link{convert.magic}};
  \code{\link{convertToDataTable}};
  \code{\link{dfDatesBetween}}; \code{\link{dfExists}};
  \code{\link{insert.column}}; \code{\link{is.not.null}};
  \code{\link{isDateBetween}};
  \code{\link{iterateDataLoad}}; \code{\link{mgsub}};
  \code{\link{pkgTest}}; \code{\link{readSQL}}
}

