% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFilteredNPSTORETData}
\alias{getFilteredNPSTORETData}
\title{getFilteredNPSTORETData}
\usage{
getFilteredNPSTORETData(df)
}
\arguments{
\item{df}{- data frame holding park data}
}
\value{
data frame - subset of park data with the following data
\itemize{
 \item Park
 \item ParkName
 \item ProjectID
 \item ProjectName
 \item StationID
 \item Station Name
 \item START_DATE
 \item QAQC_SAMPLE
 \item LocCharNameCode
 \item DISPLAY_NAME
 \item SMPL_FRAC_TYPE_NM
 \item MEDIUM
 \item DETECTION_CONDITION
 \item RESULT_TEXT
 \item UOM
 \item VALUE_STATUS
 \item VALUE_TYPE
 \item DETECTION_LIMIT
}
}
\description{
Filter NPSTORET data by standard columns & populate data frame
}
\section{Requirements}{

  Assumes the following columns exist in park dataframe
    Park, ParkName, ProjectID, ProjectName, StationID, Station Name,
    START_DATE, QAQC_SAMPLE, LocCharNameCode, DISPLAY_NAME,
    SMPL_FRAC_TYPE_NM, MEDIUM, DETECTION_CONDITION, RESULT_TEXT, UOM,
    VALUE_STATUS, VALUE_TYPE, DETECTION_LIMIT
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-05-05 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-05  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-11  \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other NPSTORET functions: \code{\link{getParkData}};
  \code{\link{getParksList}};
  \code{\link{loadNPSTORETData}};
  \code{\link{loadNPSTORETWQData}}
}

