% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIndepWQStdChars}
\alias{getIndepWQStdChars}
\title{getIndepWQStdChars}
\usage{
getIndepWQStdChars()
}
\description{
Fetch ALL NPSTORET WQ Standard Characteristics for Independent (Uncalculated) Standards
}
\section{Requirements}{

 R libraries:
  \itemize{
   \item RODBC
  }
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-06-13 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-13    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-18    \tab\tab BLC   \tab\tab Added stringsAsFactors = FALSE to prevent
                            factorization of df strings & subsequent
                            limitation of df values & fix df names & station \cr
  \tab 0.3   \tab\tab 2014-11-11    \tab\tab BLC   \tab\tab Documentation update, add check for dfRAWstds & remove sqldf depedency \cr
  \tab 0.4   \tab\tab 2015-01-06    \tab\tab BLC   \tab\tab Fix unused argument (IsDependent=0) error \cr
 }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

