% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_NPSTORET_functions.R
\name{getParkData}
\alias{getParkData}
\title{getParkData}
\usage{
getParkData(df, parkAbbrev)
}
\arguments{
\item{df}{- data frame holding park data}

\item{parkAbbrev}{- park name abrev (ARCH, CANY, etc.)}
}
\description{
Pull park data from dataframe
}
\section{Sources}{

\tabular{llllllll}{
  \tab 2014-05-07 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-05-07  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-11  \tab\tab BLC   \tab\tab Documentation update \cr
  }
}
\seealso{
Other NPSTORET functions: \code{\link{getFilteredNPSTORETData}};
  \code{\link{getParksList}};
  \code{\link{loadNPSTORETData}};
  \code{\link{loadNPSTORETWQData}}
}

