% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRAWDepWQStdChars}
\alias{getRAWDepWQStdChars}
\title{getRAWDepWQStdChars}
\usage{
getRAWDepWQStdChars()
}
\arguments{
\item{none}{}
}
\description{
fetch ALL NPSTORET WQ Standard Characteristics for Dependent (Uncalculated) Standards
}
\section{Requirements}{

 \tabular{l}{
   RODBC library package loaded \cr
   Working directory set for app \cr
   LoadNPSTORET function \cr
   sqldf library package loaded \cr
}
}

\section{Sources}{

\tabular{llllllll}{
   \tab 2014-06-25 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
   }
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-06-25    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-08    \tab\tab BLC   \tab\tab Drop unused levels \cr
  \tab 0.3   \tab\tab 2014-09-10    \tab\tab BLC   \tab\tab Remove all extra columns except CalcFormula
                            since these are time/date dependent & require lookups
                            Hardness, pH, & Temp will be added to RESULTS vs chars dataframes \cr
  \tab 0.5   \tab\tab 2014-09-18    \tab\tab BLC   \tab\tab Added stringsAsFactors = FALSE to prevent
                            factorization of df strings & subsequent
                            limitation of df values, fix spaces & reordering df \cr
 }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupDependentCharResult}},
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

