% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_WQstds_functions.R
\name{getWQStdValues}
\alias{getWQStdValues}
\title{getWQStdValues}
\usage{
getWQStdValues(dfWQstds)
}
\arguments{
\item{dfWQstds}{- NPSTORET water quality standards data frame}
}
\description{
Fetch NPSTORET WQ Standard Values for Calculated & Uncalculated Characteristics
}
\section{Requirements}{

\itemize{
 \item \code{\link[RODBC]{RODBC}}
 \item \code{\link[sqldf]{sqldf}}
}
}

\section{Sources}{

  Reference NPSTORET reportUtilities module, CalcDepStandard function
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 2014-05-21 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  \tab 2014-06-05 \tab\tab B. Campbell \tab\tab 0.2 \tab\tab Adjusted to mirror NPSTORET calculations \cr
  \tab 2014-11-13 \tab\tab B. Campbell \tab\tab 0.3 \tab\tab Documentation update \cr
  }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{lookupDependentCharResult}};
  \code{\link{lookupIndependentCharStd}}
}

