% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeEnvironment}
\alias{initializeEnvironment}
\title{initializeEnvironment}
\usage{
initializeEnvironment()
}
\description{
Load packages, functions and NPSTORET data into global (working) environment
}
\section{Requirements}{

  none
}

\section{Sources}{

\tabular{llllllll}{
  \tab 2014-09-30 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
}

\section{Revisions}{

\tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-09-30  \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-11-08  \tab\tab BLC   \tab\tab Documentation update, removed unused siny, reshape2, & stringr package calls \cr
  \tab 0.3   \tab\tab 2014-11-12  \tab\tab BLC   \tab\tab Revise SQL paths to sqlPath \cr
  }
}

\section{Notes}{

This function accomplishes the following tasks:
\itemize{
\item Increases memory limit to 4095
\item Calls necessary libraries - tools, RODBC, sqldf
\item Loads app class and packages functions into the environment
}
}
\seealso{
Other app initialization functions: \code{\link{initializeData}}
}

