% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookupDependentCharResult}
\alias{lookupDependentCharResult}
\title{lookupDependentCharResult}
\usage{
lookupDependentCharResult(depChar, park, stationID, startDate, startTimeZone,
  uom, medium, field_lab)
}
\arguments{
\item{depChar}{- NPSTORET dependent characteristic value (CharDependent)}

\item{park}{- NPSTORET park}

\item{stationID}{- NPSTORET StationID}

\item{startDate}{- NPSTORET visit StartDate}

\item{startTimeZone}{- NPSTORET visit StartTimeZone}

\item{uom}{- NPSTORET unit of measure (UOM)}

\item{medium}{- NPSTORET measurement medium (MEDIUM)}

\item{field_lab}{- NPSTORET field or lab mesurement (FIELD_LAB)}
}
\value{
resultVal - results value for the matched characteristic
}
\description{
Lookup dependent characteristic result value (pH, H20 temp, hardness)
}
\section{Requirements}{

none
}

\section{Sources}{

  \tabular{llllllll}{
  \tab 2014-07-02 \tab\tab B. Campbell \tab\tab 0.1 \tab\tab Initial version \cr
  }
  Richie Cotton, April 24, 2012
  \url{http://stackoverflow.com/questions/10294284/remove-all-special-characters-from-a-string-in-r}
  Dason, Jun 2, 2012
  \url{http://stackoverflow.com/questions/10865095/why-do-i-get-warning-longer-object-length-is-not-a-multiple-of-shorter-object-l}
}

\section{Adapted}{

  -
}

\section{Revisions}{

  \tabular{llllllll}{
  \tab 0.1   \tab\tab 2014-07-02    \tab\tab BLC   \tab\tab Initial version \cr
  \tab 0.2   \tab\tab 2014-09-11    \tab\tab BLC   \tab\tab Fixed match to return proper result value \cr
  \tab\tab\tab\tab\tab\tab\tab Replaced length with nrow to get proper # of rows (vs. df width/cols) \cr
  \tab\tab\tab\tab\tab\tab\tab Replaced == with \%in\% for comparisons \cr
  \tab 0.3   \tab\tab 2014-11-1    \tab\tab BLC   \tab\tab Documentation update \cr
  \tab 0.4   \tab\tab 2015-01-07    \tab\tab BLC   \tab\tab Add park comparator \cr
   }
}
\seealso{
Other WQ Standards functions: \code{\link{calculateStdValues}};
  \code{\link{compareToStandards}};
  \code{\link{depCharResultsLookup}};
  \code{\link{getDepCharDfs}};
  \code{\link{getDepCharResults}};
  \code{\link{getExtendedDepCharResults}};
  \code{\link{getFormulaUsed}};
  \code{\link{getHardnessLnVal}};
  \code{\link{getIndepWQStdChars}};
  \code{\link{getRAWDepWQStdChars}};
  \code{\link{getStdDepCharDisplayName}};
  \code{\link{getStdFormula}};
  \code{\link{getWQRawStdCharacteristics}};
  \code{\link{getWQStdValues}};
  \code{\link{lookupIndependentCharStd}}
}

